ingest: generalize ingest_vcfs and extract_genotypes off noodles (#138)
Closes the last hardcoded VCF assumption in the ingest path. Both the
sites-only and cohort-build CLI paths now route through FormatHandler,
so .gds and .agds inputs work the same as VCF.

Trait:
  VariantReader gains sample_names(&mut self). VcfVariantReader
  delegates to read_sample_names. GdsVariantReader's existing helper is
  now the trait impl (dropped from the inherent impl).

ingest::vcf:
  RecordContext::ingest_files takes &dyn FormatHandler and calls
  handler.open_reader instead of constructing VcfVariantReader.
  ingest_vcfs and ingest_vcfs_parallel both take a handler so they're
  no longer VCF-only despite the names. validate_headers_parallel
  becomes validate_sample_consistency, routing sample-list comparison
  through the trait. stream_genotypes takes a handler.

ingest::format:
  handler_for_format(InputFormat) returns a Box<dyn FormatHandler>.
  Single source of truth for InputFormat -> handler.

staar::genotype:
  extract_genotypes takes &dyn FormatHandler. Sample-list consistency
  check across files goes through the trait. The inherent
  read_sample_names function stays public for the STAAR command path.

store::cohort:
  CohortSources gains genotype_format. CohortHandle::build_or_load
  resolves it to a handler via handler_for_format and threads it into
  extract_genotypes.

commands/ingest:
  run_ingest dispatches Vcf and Gds the same way. No more "convert to
  VCF for now" early return. Sample-name lookup uses
  handler.open_reader(...).sample_names() instead of the VCF-only helper.

Verified locally on hsph and serial_requeue:
  cargo test --workspace -> 342 favor + 3 smoke pass, 1 ignored.
  cargo clippy --workspace --all-targets -- -D warnings -> clean.
  Production probe against
  /n/holystore01/LABS/xlin/Lab/zhouhufeng/Data/CAD/CCDG_F3/aGDSv2/
  ccdg_genome_136k_chrY.agds (13G LZMA_ra): 136,959 samples,
  14,292,827 chrY variants, variant 0 fully called (273,918 non-missing
  alleles), /chromosome[0]="Y", /allele[0]="C,T".

+++ crates/corearray-sys/tests/smoke.rs
@@ -155,11 +155,11 @@ fn probe_real_agds_from_env() {
         // dims = [n_variants, n_samples, 2]. Read one variant: pin axis 0
         // to 0, take all samples and ploidies. 2-d fallback (sample, variant).
         let (geno_count, gs, gl): (usize, Vec<i32>, Vec<i32>) = if nd == 3 {
-            let nsamp = gd[1] as i64;
-            let ploidy = gd[2] as i64;
+            let nsamp = gd[1];
+            let ploidy = gd[2];
             ((nsamp * ploidy) as usize, vec![0, 0, 0], vec![1, nsamp as i32, ploidy as i32])
         } else {
-            let nsamp = gd[0] as i64;
+            let nsamp = gd[0];
             (nsamp as usize, vec![0, 0], vec![nsamp as i32, 1])
         };
         let mut buf = vec![0i8; geno_count];
